% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{ann_return}
\alias{ann_return}
\title{Annualized return (simple)}
\usage{
ann_return(r, freq = 12)
}
\arguments{
\item{r}{Returns vector.}

\item{freq}{Periods per year (12 monthly, 252 daily).}
}
\value{
Annualized return.
}
\description{
Annualized return (simple)
}
