% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{ann_vol}
\alias{ann_vol}
\title{Annualized volatility}
\usage{
ann_vol(r, freq = 12)
}
\arguments{
\item{r}{Returns vector.}

\item{freq}{Periods per year.}
}
\value{
Annualized volatility.
}
\description{
Annualized volatility
}
