% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling.R
\name{ewma}
\alias{ewma}
\title{Exponentially weighted moving average (EWMA)}
\usage{
ewma(x, lambda = 0.94)
}
\arguments{
\item{x}{Numeric vector.}

\item{lambda}{Smoothing parameter in (0,1). Higher = smoother.}
}
\value{
EWMA series.
}
\description{
Exponentially weighted moving average (EWMA)
}
