% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{ols_table}
\alias{ols_table}
\title{OLS results table (optionally with Newey–West SE)}
\usage{
ols_table(fit, se_type = c("classic", "nw"), L = NULL)
}
\arguments{
\item{fit}{Output from ols_fit().}

\item{se_type}{"classic" or "nw"}

\item{L}{NW lag truncation if se_type = "nw"}
}
\value{
data.frame with estimate, se, t, p
}
\description{
OLS results table (optionally with Newey–West SE)
}
