% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling.R
\name{roll_mean}
\alias{roll_mean}
\title{Rolling mean}
\usage{
roll_mean(x, w)
}
\arguments{
\item{x}{Numeric vector.}

\item{w}{Window size (integer >= 1).}
}
\value{
Vector with NA for the first w-1 elements.
}
\description{
Rolling mean
}
