% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling.R
\name{roll_sd}
\alias{roll_sd}
\title{Rolling standard deviation}
\usage{
roll_sd(x, w)
}
\arguments{
\item{x}{Numeric vector.}

\item{w}{Window size.}
}
\value{
Vector with NA for the first w-1 elements.
}
\description{
Rolling standard deviation
}
