% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling.R
\name{roll_zscore}
\alias{roll_zscore}
\title{Rolling z-score}
\usage{
roll_zscore(x, w)
}
\arguments{
\item{x}{Numeric vector.}

\item{w}{Window size.}
}
\value{
Vector of z-scores using rolling mean/sd.
}
\description{
Rolling z-score
}
