% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{sharpe_ratio}
\alias{sharpe_ratio}
\title{Sharpe ratio}
\usage{
sharpe_ratio(r, rf = 0)
}
\arguments{
\item{r}{Returns vector.}

\item{rf}{Risk-free rate (scalar or vector). Default 0.}
}
\value{
Sharpe ratio (mean excess return / sd).
}
\description{
Sharpe ratio
}
