% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{yoy}
\alias{yoy}
\title{Year-over-year change}
\usage{
yoy(x, freq = 12L)
}
\arguments{
\item{x}{Numeric vector.}

\item{freq}{Seasonal frequency (12 for monthly, 4 for quarterly).}
}
\value{
YoY change with NA padding.
}
\description{
Year-over-year change
}
