% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{yoy_log}
\alias{yoy_log}
\title{Year-over-year log growth}
\usage{
yoy_log(x, freq = 12L)
}
\arguments{
\item{x}{Numeric vector (positive where computed).}

\item{freq}{Seasonal frequency (12 for monthly, 4 for quarterly).}
}
\value{
YoY log growth (in log points).
}
\description{
Year-over-year log growth
}
